+++ src/NeuralNet/Layers/Activation.php
@@ -30,7 +30,7 @@ class Activation implements Hidden
     /**
      * The width of the layer.
      *
-     * @var int<0,max>|null
+     * @var positive-int|null
      */
     protected ?int $width = null;
 
@@ -62,7 +62,7 @@ class Activation implements Hidden
      * @internal
      *
      * @throws \Rubix\ML\Exceptions\RuntimeException
-     * @return int<0,max>
+     * @return positive-int
      */
     public function width() : int
     {
@@ -79,8 +79,8 @@ class Activation implements Hidden
      *
      * @internal
      *
-     * @param int<0,max> $fanIn
-     * @return int<0,max>
+     * @param positive-int $fanIn
+     * @return positive-int
      */
     public function initialize(int $fanIn) : int
     {