+++ src/NeuralNet/Layers/Continuous.php
@@ -48,7 +48,7 @@ class Continuous implements Output
     /**
      * Return the width of the layer.
      *
-     * @return int<0,max>
+     * @return positive-int
      */
     public function width() : int
     {
@@ -59,9 +59,9 @@ class Continuous implements Output
      * Initialize the layer with the fan in from the previous layer and return
      * the fan out for this layer.
      *
-     * @param int<0,max> $fanIn
+     * @param positive-int $fanIn
      * @throws \Rubix\ML\Exceptions\InvalidArgumentException
-     * @return int<0,max>
+     * @return positive-int
      */
     public function initialize(int $fanIn) : int
     {