+++ src/NeuralNet/Layers/Dropout.php
@@ -43,7 +43,7 @@ class Dropout implements Hidden
     /**
      * The width of the layer.
      *
-     * @var int<0,max>|null
+     * @var positive-int|null
      */
     protected ?int $width = null;
 
@@ -75,7 +75,7 @@ class Dropout implements Hidden
      * @internal
      *
      * @throws \Rubix\ML\Exceptions\RuntimeException
-     * @return int<0,max>
+     * @return positive-int
      */
     public function width() : int
     {
@@ -92,8 +92,8 @@ class Dropout implements Hidden
      *
      * @internal
      *
-     * @param int<0,max> $fanIn
-     * @return int<0,max>
+     * @param positive-int $fanIn
+     * @return positive-int
      */
     public function initialize(int $fanIn) : int
     {