+++ src/NeuralNet/Layers/Multiclass.php
@@ -87,20 +87,20 @@ class Multiclass implements Output
     /**
      * Return the width of the layer.
      *
-     * @return int<0,max>
+     * @return positive-int
      */
     public function width() : int
     {
-        return count($this->classes);
+        return max(1, count($this->classes));
     }
 
     /**
      * Initialize the layer with the fan in from the previous layer and return
      * the fan out for this layer.
      *
-     * @param int<0,max> $fanIn
+     * @param positive-int $fanIn
      * @throws \Rubix\ML\Exceptions\InvalidArgumentException
-     * @return int<0,max>
+     * @return positive-int
      */
     public function initialize(int $fanIn) : int
     {
@@ -199,10 +199,10 @@ class Multiclass implements Output
                 ->divide($output->n());
         }
 
-        $dL = $this->costFn->differentiate($output, $expected)
+        $dLoss = $this->costFn->differentiate($output, $expected)
             ->divide($output->n());
 
         return $this->softmax->differentiate($input, $output)
-            ->multiply($dL);
+            ->multiply($dLoss);
     }
 }