+++ tests/NeuralNet/Layers/ActivationTest.php
@@ -18,7 +18,7 @@ use PHPUnit\Framework\TestCase;
 class ActivationTest extends TestCase
 {
     /**
-     * @var int<0,max>
+     * @var positive-int
      */
     protected $fanIn;