+++ tests/NeuralNet/Layers/DropoutTest.php
@@ -19,7 +19,7 @@ class DropoutTest extends TestCase
     protected const RANDOM_SEED = 0;
 
     /**
-     * @var int<0,max>
+     * @var positive-int
      */
     protected $fanIn;