+++ tests/NeuralNet/Layers/PReLUTest.php
@@ -21,7 +21,7 @@ class PReLUTest extends TestCase
     protected const RANDOM_SEED = 0;
 
     /**
-     * @var int<0,max>
+     * @var positive-int
      */
     protected $fanIn;