+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+- 1.3.2
+    - Optimize Binary output layer
+
 - 1.3.1
     - Update to Ok Bloomer 1.0 stable